+++ test/test.js
@@ -19,7 +19,6 @@ describe('i18next', function() {
       lowerCaseLng: false,
       ns: 'translation',
       fallbackToDefaultNS: false,
-      fallbackOnNull: true,
       resGetPath: 'locales/__lng__/__ns__.json',
       dynamicLoad: false,
       useLocalStorage: false,
@@ -33,7 +32,8 @@ describe('i18next', function() {
       parseMissingKey: '',
       interpolationPrefix: '__',
       interpolationSuffix: '__',
-      shortcutFunction: 'sprintf'
+      shortcutFunction: 'sprintf',
+      objectTreeKeyHandler: null
     };
   });
 
@@ -988,8 +988,56 @@ describe('i18next', function() {
     
     });
   
+    describe('using objectTreeKeyHandler', function() {
+    
+      beforeEach(function(done) {
+        i18n.init(i18n.functions.extend(opts, {
+          objectTreeKeyHandler: function(key, value, lng, ns, opts) {
+            return i18n.t(key + '.a');
+          },
+          resStore: {
+            'en-US': { translation: { 'simpleTest': { a: 'a value', b: 'b value' } } }
+          },
+          returnObjectTrees: false
+        }), function(t) { done(); } );
+      });
+    
+      it('it should apply objectTreeKeyHandler', function() {
+        expect(i18n.t('simpleTest')).to.be('a value');
+      });
+    
+    });
+  
   });
   describe('translation functionality', function() {
+  
+    describe('keys with non supported values', function() {
+    
+      var resStore = {
+        dev: { translation: {  } },
+        en: { translation: {  } },            
+        'en-US': { 
+          translation: {                      
+            test: 'hi'
+          } 
+        }
+      };
+      
+      beforeEach(function(done) {
+        i18n.init(i18n.functions.extend(opts, { resStore: resStore }),
+          function(t) { done(); });
+      });
+    
+    
+      it('it should not break on null key', function() {console.log('here', i18n.t(null))
+        expect(i18n.t(null)).to.be('');
+      });
+    
+      it('it should not break on undefined key', function() {
+        expect(i18n.t(undefined)).to.be('');
+      });
+      
+    });
   
     describe('resource is missing', function() {
       var resStore = {
@@ -2135,6 +2183,50 @@ describe('i18next', function() {
         it('it should set inner html', function() {
           $('#container').i18n();
           expect($('#inner').html()).to.be('test');
+        });
+        
+      });
+    
+      describe('extended - append html', function() {
+    
+        var resStore = {
+          dev: { translation: {  } },
+          en: { translation: {  } },            
+          'en-US': { translation: { 'simpleTest': '<div id="inner">test</div>' } }
+        };
+        
+        beforeEach(function(done) {
+          setFixtures('<div id="container" data-i18n="[append]simpleTest"><div></div></div>');
+    
+          i18n.init(i18n.functions.extend(opts, { resStore: resStore }),
+            function(t) {  done(); });
+        });
+    
+        it('it should append html', function() {
+          $('#container').i18n();
+          expect($('#container').html()).to.be('<div></div><div id="inner">test</div>');
+        });
+        
+      });
+    
+      describe('extended - prepend html', function() {
+    
+        var resStore = {
+          dev: { translation: {  } },
+          en: { translation: {  } },            
+          'en-US': { translation: { 'simpleTest': '<div id="inner">test</div>' } }
+        };
+        
+        beforeEach(function(done) {
+          setFixtures('<div id="container" data-i18n="[prepend]simpleTest"><div></div></div>');
+    
+          i18n.init(i18n.functions.extend(opts, { resStore: resStore }),
+            function(t) {  done(); });
+        });
+    
+        it('it should append html', function() {
+          $('#container').i18n();
+          expect($('#container').html()).to.be('<div id="inner">test</div><div></div>');
         });
         
       });